feat: initial parser implementation
The parser doesn't yet fully conform to the spec, but it's in a usable
state. There are a few tests in place, but more are needed to make sure
all edge-cases are covered.

The spec isn't stable yet, so this library will remain at a non-stable
version until the spec moves to 1.0, after which this library will
follow.