refactor: improve string parsing
This changes introduces the usage of [Nom] as the parser combinator
library used to construct and validate a conventional commit based on a
provided string.

The initial implementation takes care to keep the same rules as the old
one, but the goal is to iterate on this implementation to better
represent the actual specification, and tidy up some edge-cases.

Because of this change, the `Commit` struct now no longer owns the
string from which the commit was created, but instead borrows slices
into the string representing the different components of the conventional
commit,

[Nom]: https://docs.rs/nom

BREAKING CHANGE: Because `Commit` now holds a reference, it can no
longer implement [`FromStr`]. Instead, there is now a `Commit::new`
function with the same signature. When upgrading, you'll need to rename
all `Commit::from_str` references to `Commit::new`.

[`FromStr`]: https://doc.rust-lang.org/std/str/trait.FromStr.html