feat: error struct wrapping error kind
This is a first step towards better error handling with user-friendly
error messages and programmatic access to error spans for other
libraries.

BREAKING CHANGE: The name of the error type is the same, but it is
switched from an enum to a struct. You can still access the original
enum using the `kind` field on the struct. To upgrade, match over
`err.kind` instead of `err`.