feat: remove lifetime from Error object
While borrowing the commit as a string slice is more efficient, the
ergonomics hit of having to propagate the lifetime on the error object
up the chain on all objects dealing with the error is too much to
justify the need of the lifetime.

We can circle back to this if there is a real need for this, but I doubt
it.