feat: support compound noun types/scopes
The spec defines:

> Commits MUST be prefixed with a type, which consists of a noun

and:

> A scope MUST consist of a noun

But especially for scopes, this can become very restrictive if you have
parts of your code that require two or more words to define the noun of
the subject.

To keep things simple, and go with the least restrictive option, both
types and scopes now support compounded nouns by supporting both spaces
and dashes in them.