feat: improve type and scope parser logic
Types and scopes are now properly validated to be compound nouns.

Specifically, this means they have to start and end with an alphanumeric
character, and can have optional spaces and hyphens in-between.

It could be further improved to reject two or more spaces or hyphens
next to each other, but this will do as a first step.