refactor!: rename "trailer" to "footer"
The first version of the spec now includes support for Git trailers, but
calls them "footers" instead[1].

The conventional commit footers are very similar to Git trailers, except
that they only support two separator types (`:<space>` and `<space>#`)
and does not require whitespace at the start of a line.

This change renames all references of "trailer" to "footer", and "key"
to "token", but leaves a few comments referencing the similarity between
footers and trailers.