feat!: mark "BREAKING-CHANGE" commits as breaking
Currently a commit with a footer token "BREAKING CHANGE" is marked as a
breaking commit (similar to a commit with a bang (`!`) in the header).

With this change, as per the spec, a "BREAKING-CHANGE" footer token also
marks the commit as breaking.

Additionally, the parser now correctly marks the first footer as a
footer, if no body is provided.