fix: prefer description over BREAKING CHANGE
The conventional commit specification says the following about breaking
changes indicated with ! after the type and BREAKING CHANGE footers:

> If included in the type/scope prefix, breaking changes MUST be
  indicated by a ! immediately before the :. If ! is used, BREAKING
  CHANGE: MAY be omitted from the footer section, and the commit
  description SHALL be used to describe the breaking change.

This updates the commit parser to prefer the description from the
subject line over the value of the BREAKING CHANGE footer when both
describe breaking changes. This is more inline with the specification
and also allows consumers to enumerate all breaking change descriptions
in the case where the subject and footers are used to list more than one
breaking change.